requirements.txt
python-dotenv
pandas
numpy
seaborn
pymongo
certifi
pymongo[srv]


##-e .